Server.js modified to pass user info from API. Middleware created for login and auth. Database middleware updated to accept LinkedIn values
